Pydantic структуры перенесены в отдельный файл
